Move Tika substitutions to proper package
Also add a comment about removing the PDFBox one when migrating to
GraalVM 21.
