+++ extensions/tika/runtime/src/main/java/io/quarkus/tika/runtime/graal/CleanerNotSupportedSubstitution.java
@@ -1,4 +1,4 @@
-package io.quarkus.tika.graalvm;
+package io.quarkus.tika.runtime.graal;
 
 @com.oracle.svm.core.annotate.Substitute
 @com.oracle.svm.core.annotate.TargetClass(className = "org.apache.poi.poifs.nio.CleanerUtil")
